Feature/149/v2 features

+++ app/scripts/clear_cache.js
@@ -1,24 +1,23 @@
-clearRailsCache();
+const https = require('https');
 
+export default function clearRailsCache() {
+  if (!process.env.RAIL_SERVERS) { return; }
+  const serverUrls = process.env.RAIL_SERVERS.split(',');
 
-function clearRailsCache() {
-  var server_urls = process.env.RAIL_SERVERS.split(',');
-  var https = require('https');
-
-  server_urls.forEach(function(url) {
-    var options = {
+  serverUrls.forEach((url) => {
+    const options = {
       host: url,
       port: 443,
       auth: 'fandpstaging:auth4fandpstaging',
       path: '/api/v1/rails_cache?systems_key=f3UwF9ftw',
-      method: 'DELETE'
-    }
+      method: 'DELETE',
+    };
 
-    var req = https.request(options, function(res) {
-      console.warn('Cache Clear Status: ' + res.statusCode);
+    const req = https.request(options, (res) => {
+      console.warn(`Cache Clear Status: ${res.statusCode}`);
     });
     req.end();
-    req.on('error', function(e) {
+    req.on('error', (e) => {
       console.error(e);
     });
   });
